Show a different starting message when creating Theme App Extensions
Fixes #168

In the context of an `extension create` command, we don't know what type
of project is being created until after the AskType question has been
asked.

In order to provide project-type specific messages in the `create`
command, we load messages via the specification handler, which knows the
project type, and can access the type specific message overrides.