Add customization of the `:BindAddress` web server option
When running in some container environments, binding to `127.0.0.1` will
not actually allow the host machine to properly access the application
through standard port-forwarding. In most cases, this can be resolved by
having the web server bind to `0.0.0.0` instead. But there could be
other reasons why other development environments might need something
other than those two, so we'll add a simple option that can be passed
into `shopify theme serve` that will adjust that address binding while
keeping the default as `127.0.0.1`.

Note: you can typically still access the development server using
`127.0.0.1` even if you set the binding address to be something else,
because on the host machine the addresses for `127.0.0.1` and
`localhost` typically go to the same interface which will have
port-forwarding enabled to the container instance which will then
communicate properly with the host.