Fix app serve after rails update
For whatever reason, it seems that the wrong rails is found if the
versions of Rails in the homebrew-installed CLI vs. your app have gotten
out of sync, at least in certain environments.

Part of the issue seems to be that the gem path isn't passed through,
and instead ends up pointing to the homebrew-installed gem path, but
that won't be the same as the gems you've bundle installed in your
Rails app!